requirements.txt
fastapi
uvicorn
fastapi-cache2[redis]